opt: refactor inverted index code into new invertedidx package
This commit creates a new package, invertedidx, and refactors some of the
code in xform that was dealing with inverted indexes into this package.
This new package is analagous to the idxconstraint package: In the same way
that the code in the idxconstraint package uses the constraint package for
constraining forward indexes, invertedidx uses invertedexpr for constraining
inverted indexes.

Release note: None
